Move reclaiming code to utils.reclaim
They don't belong into the `legacy` module. Even though I'm not sure
what makes the rest of the functions there "legacy".
