+++ scenario_player/utils/__init__.py
@@ -3,11 +3,8 @@ from scenario_player.utils.legacy import (
     DummyStream,
     TimeOutHTTPAdapter,
     post_task_state_to_rc,
-    reclaim_erc20,
-    reclaim_eth,
     send_rc_message,
     wait_for_txs,
-    withdraw_from_udc,
 )
 
 __all__ = [
@@ -15,9 +12,6 @@ __all__ = [
     "ConcatenableNone",
     "DummyStream",
     "wait_for_txs",
-    "reclaim_eth",
-    "reclaim_erc20",
-    "withdraw_from_udc",
     "post_task_state_to_rc",
     "send_rc_message",
 ]
